+++ packages/vscode-extension/src/scopes.ts
@@ -30,7 +30,7 @@ export type ScopeClassification =
 type Api = {
 	title: string;
 	version: string | number | null;
-	documentationLink: string;
+	documentationLink?: string;
 };
 
 type Scope = {
@@ -41,12 +41,8 @@ type Scope = {
 
 export const SCOPES = new Map<string, Scope>();
 
-for (const { title, version, documentationLink, scopes } of GOOGLE_APIS) {
-	for (const { id, description } of scopes) {
-		console.log(
-			`Processing scope: ${id} - ${description} (${title} v${version})`,
-		);
-
+for (const { title, version, documentationLink, scopes } of GOOGLE_APIS || []) {
+	for (const { id, description } of scopes || []) {
 		if (!SCOPES.has(id)) {
 			SCOPES.set(id, {
 				description: description,
@@ -66,6 +62,50 @@ for (const { title, version, documentationLink, scopes } of GOOGLE_APIS) {
 			documentationLink,
 		});
 	}
+}
+
+export const SCRIPT_EXTERNAL_REQUEST_SCOPE =
+	"https://www.googleapis.com/auth/script.external_request";
+SCOPES.set(SCRIPT_EXTERNAL_REQUEST_SCOPE, {
+	description: "Connect to an external service",
+	apis: [],
+});
+
+export const SCRIPT_CONTAINER_UI_SCOPE =
+	"https://www.googleapis.com/auth/script.container.ui";
+SCOPES.set(SCRIPT_CONTAINER_UI_SCOPE, {
+	description:
+		"Display and run third-party web content in prompts and sidebars inside Google applications.",
+	apis: [],
+});
+
+export const SCRIPT_SEND_MAIL_SCOPE =
+	"https://www.googleapis.com/auth/script.send_mail";
+SCOPES.set(SCRIPT_SEND_MAIL_SCOPE, {
+	description: "Send email on your behalf.",
+	apis: [],
+});
+
+/**
+ * Scopes that are only available for the current document.
+ *
+ * @see https://developers.google.com/workspace/add-ons/concepts/workspace-scopes#editor-scopes
+ * @see https://justin.poehnelt.com/posts/apps-script-currentonly-scopes/
+ */
+export const CURRENT_ONLY_SCOPES = [
+	"https://www.googleapis.com/auth/documents.currentonly",
+	"https://www.googleapis.com/auth/forms.currentonly",
+	"https://www.googleapis.com/auth/presentations.currentonly",
+	"https://www.googleapis.com/auth/spreadsheets.currentonly",
+];
+
+for (const scope of CURRENT_ONLY_SCOPES) {
+	SCOPES.set(scope, {
+		description:
+			"Access the current document, sheet, presentation, or form. The `currentonly` scope is only available within Apps Script Services. This does not include Apps Script Advanced Services or direct calls to Google Workspace APIs.",
+		classification: ScopeClassification.NON_SENSITIVE,
+		apis: [],
+	});
 }
 
 const RESTRICTED_SCOPES = [