+++ packages/vscode-extension/tsup.config.mjs
@@ -1,7 +1,11 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-	entry: ["src/extension.ts", "src/test/**/*.ts"],
+	entry: [
+		"src/extension.ts",
+		"src/test/scopes.test.ts",
+		"src/test/completion.test.ts",
+	],
 	outDir: "out",
 	clean: true,
 	external: ["vscode"],